CRUD items only (oauth2 implemented)

+++ vagrant/catalog/item-catalog.py
@@ -69,85 +69,39 @@ def showCategory(category_id):
     Base.metadata.bind = engine
     DBSession = sessionmaker(bind=engine)
     session = DBSession()
-
     category = session.query(Category).filter_by(id=category_id).one()
     items = session.query(Item).filter_by(category_id=category.id).all()
     itemssize = len(items)
-
-
-    return render_template('category.html', category = category, items = items, itemssize=itemssize)
-
-# 3) New category
-@app.route('/categories/new/', methods=['GET', 'POST'])
-def newCategory():
-    # POST method add the new category then go to homepage
-    engine = create_engine('sqlite:///itemcatalog.db')
-    Base.metadata.bind = engine
-    DBSession = sessionmaker(bind=engine)
-    session = DBSession()
-
-    if request.method == 'POST':
-        if ('username' not in login_session):
-            return redirect('/')
+    # check if user is authenticated
+    if ('username' not in login_session):
+        login = 0
+    else:
+        if login_session['username'] == None:
+            login = 0
         else:
-            if login_session['username'] == None:
-                return redirect('/')
-            else:
-                user = session.query(User).filter_by(name = login_session['username']).one()
-                newCategory = Category(name = ""+request.form['name'], user = user)
-                session.add(newCategory)
-                session.commit()
-                return redirect(url_for('showCategories'))
-    # GET method show add new category template
-    else:
-        return render_template('category-new.html')
+            login = 1
+    return render_template('category.html', category = category, items = items, itemssize=itemssize, login=login)
 
-
-# 4) Edit category
-@app.route('/categories/<int:category_id>/edit/', methods=['GET', 'POST'])
-def editCategory(category_id):
-    editcategory = session.query(Category).filter_by(id = category_id).one()
-    # POST edit the category with the form given and return to homepage
-    if request.method == 'POST':
-        #if (editcategory.user.email == login_session['email']) and (editcategory.user.name == login_session['name']):
-        editcategory.name = ""+request.form['name']
-        session.add(editcategory)
-        session.commit()
-        # flash("Category edited")
-        return redirect(url_for('showCategories'))
-    # GET pass category to edit category template
-    else:
-        return render_template('category-edit.html', category = editcategory)
-
-# 5) Delete category
-@app.route('/categories/<int:category_id>/delete/', methods=['GET', 'POST'])
-def deleteCategory(category_id):
-    deletecategory = session.query(Category).filter_by(id = category_id).one()
-    # you must delete items of that category
-    categoryitems = session.query(Item).filter_by(category_id = category_id).all()
-    # POST delete that category and return to homepage
-    if request.method == 'POST':
-        # delete items of that category
-        for ci in categoryitems:
-            session.delete(ci)
-            session.commit()
-        session.delete(deletecategory)
-        session.commit()
-        # flash("Category deleted")
-        return redirect(url_for('showCategories'))
-    # GET pass category to delete category template
-    else:
-        return render_template('category-delete.html', category = deletecategory)
-
-# 6) Item page
+# 3) Item page
 @app.route('/categories/<int:category_id>/items/<int:item_id>/')
 def showItem(category_id, item_id):
     # find item descriptionpage
     category = session.query(Category).filter_by(id=category_id).one()
     item = session.query(Item).filter_by(id=item_id).one()
-    return render_template('item.html', category = category, item = item)
+    # check if user is item creator
+    if ('username' not in login_session):
+        creator = 0
+    else:
+        if login_session['username'] == None:
+            creator = 0
+        else:
+            if (item.user.name == login_session['username']) and (item.user.email == login_session['email']):
+                creator = 1
+            else:
+                creator = 0
+    return render_template('item.html', category = category, item = item, creator=creator)
 
-# 7) New item
+# 4) New item
 @app.route('/categories/<int:category_id>/items/new/', methods=['GET', 'POST'])
 def newItem(category_id):
     # POST method add the new item then go to homepage
@@ -156,55 +110,87 @@ def newItem(category_id):
     Base.metadata.bind = engine
     DBSession = sessionmaker(bind=engine)
     session = DBSession()
-
     if request.method == 'POST':
-        category = session.query(Category).filter_by(id = category_id).one()
-        user = session.query(User).filter_by(id = category.user_id).one()
-        newItem = Item(name = ""+request.form['name'], description = ""+request.form['description'], user = user, category = category, created_at = datetime.datetime.now())
-        session.add(newItem)
-        session.commit()
-        # flash("Category added")
-        return redirect(url_for('showCategory', category_id=category_id))
+        # only add item if user is authenticated
+        if ('username' not in login_session):
+            # flash("Must be logged in to add items")
+            return redirect(url_for('showCategory', category_id=category_id))
+        else:
+            if login_session['username'] == None:
+                # flash("Must be logged in to add items")
+                return redirect(url_for('showCategory', category_id=category_id))
+            else:
+                category = session.query(Category).filter_by(id = category_id).one()
+                user = session.query(User).filter_by(email = login_session['email']).one()
+                newItem = Item(name = ""+request.form['name'], description = ""+request.form['description'], user = user, category = category, created_at = datetime.datetime.now())
+                session.add(newItem)
+                session.commit()
+                return redirect(url_for('showCategory', category_id=category_id))
     # GET method show add new category template
     else:
         category = session.query(Category).filter_by(id = category_id).one()
         return render_template('item-new.html', category = category)
 
-# 8) Edit item
+# 5) Edit item
 @app.route('/categories/<int:category_id>/items/<int:item_id>/edit/', methods=['GET', 'POST'])
 def editItem(category_id, item_id):
     edititem = session.query(Item).filter_by(id = item_id).one()
     # POST edit the item with the form given and return to category
     if request.method == 'POST':
-        category = session.query(Category).filter_by(name = ""+request.form['categorybox']).one()
-        edititem.name = ""+request.form['name']
-        edititem.description = ""+request.form['description']
-        edititem.category = category
-        session.add(edititem)
-        session.commit()
-        # flash("Category edited")
-        return redirect(url_for('showItem', category_id=category_id, item_id=item_id ))
+        if ('username' not in login_session):
+            # flash("Must be logged in and item owner to edit item")
+            return redirect(url_for('showCategory', category_id=category_id))
+        else:
+            if login_session['username'] == None:
+                # flash("Must be logged in and item owner to edit item")
+                return redirect(url_for('showCategory', category_id=category_id))
+            else:
+                if (edititem.user.name == login_session['username']) and (edititem.user.email == login_session['email']):
+                    category = session.query(Category).filter_by(name = ""+request.form['categorybox']).one()
+                    edititem.name = ""+request.form['name']
+                    edititem.description = ""+request.form['description']
+                    edititem.category = category
+                    session.add(edititem)
+                    session.commit()
+                    # flash("Item edited")
+                    return redirect(url_for('showItem', category_id=category_id, item_id=item_id ))
+                else:
+                    # flash("Must be logged in and item owner to edit item")
+                    return redirect(url_for('showCategory', category_id=category_id))
     # GET pass item to edit item template
     else:
         category1 = session.query(Category).filter_by(id = category_id).one()
         categories = session.query(Category).all()
         return render_template('item-edit.html', item = edititem, category=category1, categories = categories)
 
-# 9) Delete item
+# 6) Delete item
 @app.route('/categories/<int:category_id>/items/<int:item_id>/delete/', methods=['GET', 'POST'])
 def deleteItem(category_id, item_id):
     deleteitem = session.query(Item).filter_by(id = item_id).one()
     # POST delete that item and return to homepage
     if request.method == 'POST':
-        session.delete(deleteitem)
-        session.commit()
-        # flash("Item deleted")
-        return redirect(url_for('showCategory', category_id=category_id))
+        if ('username' not in login_session):
+            # flash("Must be logged in and item owner to delete item")
+            return redirect(url_for('showCategory', category_id=category_id))
+        else:
+            if login_session['username'] == None:
+                # flash("Must be logged in and item owner to delete item")
+                return redirect(url_for('showCategory', category_id=category_id))
+            else:
+                if (deleteitem.user.name == login_session['username']) and (deleteitem.user.email == login_session['email']):
+                    session.delete(deleteitem)
+                    session.commit()
+                    # flash("Item deleted")
+                    return redirect(url_for('showCategory', category_id=category_id))
+                else:
+                    # flash("Must be logged in and item owner to delete item")
+                    return redirect(url_for('showCategory', category_id=category_id))
+
     # GET pass item to delete item template
     else:
         return render_template('item-delete.html', item = deleteitem, category_id=category_id)
 
-# 10) JSON endpoint
+# 7) JSON endpoint
 @app.route('/catalog/JSON/')
 def catalogJSON():
     engine = create_engine('sqlite:///itemcatalog.db')
@@ -223,7 +209,7 @@ def catalogJSON():
         c['items'] = Items
     return jsonify(Categories=Categories)
 
-# 1) Show users
+# 8) Show users
 @app.route('/users/')
 def showUsers():
     # find all users to pass to html template
